moved Exception to exception package
