+++ lib-core/src/main/java/ch/hsr/adv/lib/core/logic/ADVCore.java
@@ -4,6 +4,7 @@ import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.commons.core.logic.util.ADVException;
 import ch.hsr.adv.lib.core.access.Connector;
 import ch.hsr.adv.lib.core.logic.domain.Session;
+import ch.hsr.adv.lib.core.logic.exceptions.ADVConnectionException;
 import ch.hsr.adv.lib.core.logic.util.*;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
